# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/icps/{icpId}/qualify": {
      "post": {
        "operationId": "LeadController_qualify_v1",
        "parameters": [
          {
            "name": "icpId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "accepted",
                        "eventId"
                      ],
                      "properties": {
                        "accepted": {
                          "type": "boolean"
                        },
                        "eventId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "对 ACTIVE ICP 的全部候选公司做六维评分 → Lead + 四队列（异步）",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:write"
        ]
      }
    },
    "/api/v1/leads": {
      "get": {
        "operationId": "LeadController_list_v1",
        "parameters": [
          {
            "name": "icpId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queue",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "Lead（六维分+队列+裁决状态）"
                      }
                    },
                    "page": {
                      "type": "object",
                      "required": [
                        "next_cursor",
                        "has_more"
                      ],
                      "properties": {
                        "next_cursor": {
                          "type": "string",
                          "nullable": true,
                          "description": "下一页游标；null = 没有更多"
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lead 列表（?icpId=&queue=recommended|needs_review|rejected|suppressed，按分数排序）",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/icps/{icpId}/lead-queues": {
      "get": {
        "operationId": "LeadController_queues_v1",
        "parameters": [
          {
            "name": "icpId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "recommended",
                        "needs_review",
                        "rejected",
                        "suppressed"
                      ],
                      "properties": {
                        "recommended": {
                          "type": "integer"
                        },
                        "needs_review": {
                          "type": "integer"
                        },
                        "rejected": {
                          "type": "integer"
                        },
                        "suppressed": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "四队列计数（推荐/待确认/拒绝/禁联）",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/leads/{leadId}": {
      "get": {
        "operationId": "LeadController_get_v1",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Lead（六维分+队列+裁决状态）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lead 详情：六维分 + 规则逐条评估依据 + 公司/联系人 + 裁决历史",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:read",
          "personal-data:read"
        ]
      }
    },
    "/api/v1/leads/{leadId}/accept": {
      "post": {
        "operationId": "LeadController_accept_v1",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Lead（六维分+队列+裁决状态）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "接受 Lead（→ QUALIFIED，发 LeadQualified —— 交给 Campaign 的出口）",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:review"
        ]
      }
    },
    "/api/v1/leads/{leadId}/reject": {
      "post": {
        "operationId": "LeadController_reject_v1",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Lead（六维分+队列+裁决状态）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "拒绝 Lead（→ REJECTED，原因留痕做质量反馈）",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:review"
        ]
      }
    },
    "/api/v1/leads/{leadId}/sanctions-review": {
      "post": {
        "operationId": "LeadController_sanctionsReview_v1",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanctionsReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "leadId": {
                          "type": "string"
                        },
                        "reviewState": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "制裁筛查复核裁决（第五门人审）：误报清白 → 回落队列；真命中确认 → 留隔离，永不交付",
        "tags": [
          "Leads"
        ],
        "x-required-scopes": [
          "acquisition:review",
          "compliance:manage"
        ]
      }
    },
...
  },
  "info": {
    "title": "Global API",
    "description": "出海企业 AI 全球客户开发与增长平台 · 后端 API（前端接入见 packages/contracts/INTEGRATION.md）",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "同源部署（相对路径；具体 host 由部署环境决定）"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AcceptLeadDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "RejectLeadDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "拒绝原因（回流做评分质量反馈）"
          }
        }
      },
      "SanctionsReviewDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "cleared_false_positive",
              "confirmed_true_hit"
            ],
            "description": "复核裁决：误报清白（回落队列，抑制复发）/ 真命中确认（留隔离，永不交付）"
          },
          "note": {
            "type": "string",
            "description": "复核备注（留痕）"
          }
        },
        "required": [
          "decision"
        ]
      },
...
    }
  }
}